+++ src/components/DataTable/selectColumnDef.tsx
@@ -1,12 +1,13 @@
 import { Checkbox } from '@/components/Checkbox/Checkbox'
 import type { ColumnDef } from '@tanstack/react-table'
+import type { DataTableFeatures } from './features'
 
-export const selectColumnDef: ColumnDef<unknown> = {
+export const selectColumnDef: ColumnDef<DataTableFeatures, Record<string, unknown>> = {
   id: 'select',
   header: ({ table }) => (
     <Checkbox
       checked={table.getIsAllPageRowsSelected()}
-      indeterminate={table.getIsSomeRowsSelected()}
+      indeterminate={table.getIsSomePageRowsSelected() && !table.getIsAllPageRowsSelected()}
       onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
       aria-label='Select all'
     />
@@ -14,10 +15,11 @@ export const selectColumnDef: ColumnDef<unknown> = {
   cell: ({ row }) => (
     <Checkbox
       checked={row.getIsSelected()}
-      onCheckedChange={row.getToggleSelectedHandler()}
+      onCheckedChange={(value) => row.toggleSelected(!!value)}
       aria-label='Select row'
     />
   ),
   size: 10,
   minSize: 10,
+  enableResizing: false,
 } as const